add logout, signup, Gooogle authentication
