+++ src/app/user/login/page.tsx
@@ -1,79 +1,38 @@
 "use client";
 
-import Head from "next/head";
+import EmailAuth from "@/components/auth/email/EmailAuth";
+import { Auth } from "@supabase/auth-ui-react";
+import { ThemeSupa } from "@supabase/auth-ui-shared";
 import { supabase } from "@/libs/supabase";
-import { useState } from "react";
+
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 
 export default function Login() {
   const router = useRouter();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [passwordConf, setPasswordConf] = useState("");
-
-  const onLogin = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    try {
-      const { error: signInError } = await supabase.auth.signInWithPassword({
-        email: email,
-        password: password,
-      });
-      if (signInError) {
-        throw signInError;
-      }
-      await router.push("/");
-    } catch {
-      alert("エラーが発生しました");
-    }
-  };
-
   return (
     <div>
-      <Head>
-        <title>ログイン画面</title>
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
+      <h1>Login</h1>
       <main>
+        <EmailAuth />
+        <Auth
+          supabaseClient={supabase}
+          appearance={{ theme: ThemeSupa }}
+          providers={["google", "github"]}
+          onlyThirdPartyProviders
+          localization={{
+            variables: {
+              sign_in: {
+                social_provider_text: "{{provider}}でログイン",
+              },
+            },
+          }}
+          redirectTo="http://localhost:3000/auth/callback"
+        />
         <div>
-          <form onSubmit={onLogin}>
-            <div>
-              <label>メールアドレス</label>
-              <input
-                type="email"
-                required
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-              />
-            </div>
-            <div>
-              <label>パスワード</label>
-              <input
-                type="password"
-                required
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-              />
-            </div>
-            <div>
-              <label>パスワード（確認）</label>
-              <input
-                type="password"
-                required
-                value={passwordConf}
-                onChange={(e) => setPasswordConf(e.target.value)}
-              />
-            </div>
-            <div>
-              <button type="submit">ログイン</button>
-              <br />
-              <Link href="signup">
-                ユーザー登録がお済みでない方はこちらから
-              </Link>
-              <br />
-              <Link href="sendEmail">パスワードをお忘れの方はこちらから</Link>
-            </div>
-          </form>
+          <Link href="signup">ユーザー登録がお済みでない方はこちらから</Link>
+          <br />
+          <Link href="sendEmail">パスワードをお忘れの方はこちらから</Link>
         </div>
       </main>
       <footer></footer>
